# docs/swagger.yaml
definitions:
...
  errors.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  features.createFeatureRequest:
    properties:
      config:
        additionalProperties: {}
        type: object
      created_by:
        type: string
      description:
        type: string
      name:
        type: string
      slug:
        type: string
      type:
        enum:
        - static
        - metered
        type: string
    required:
    - created_by
    - name
    - slug
    - type
    type: object
  features.updateFeatureRequest:
    properties:
      config:
        additionalProperties: {}
        type: object
      description:
        maxLength: 255
        minLength: 10
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      updated_by:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - updated_by
    type: object
  meters.createMeterRequest:
    properties:
      aggregation:
        enum:
        - count
        - sum
        - avg
        - unique_count
        - min
        - max
        type: string
      created_by:
        type: string
      description:
        type: string
      event_type:
        type: string
      name:
        type: string
      populate:
        type: boolean
      properties:
        items:
          type: string
        minItems: 1
        type: array
      slug:
        type: string
      value_property:
        type: string
    required:
    - aggregation
    - created_by
    - event_type
    - name
    - populate
    - properties
    - slug
    type: object
...
  meters.updateMeterRequest:
    properties:
      description:
        maxLength: 255
        minLength: 3
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      updated_by:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - updated_by
    type: object
  models.AggregationEnum:
    enum:
    - count
    - sum
    - avg
    - unique_count
    - min
    - max
    type: string
    x-enum-varnames:
    - AggregationCount
    - AggregationSum
    - AggregationAvg
    - AggregationUniqueCount
    - AggregationMin
    - AggregationMax
...
  models.Feature:
    properties:
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      tenant_slug:
        type: string
      type:
        $ref: '#/definitions/models.FeatureTypeEnum'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.FeatureTypeEnum:
    enum:
    - static
    - metered
    type: string
    x-enum-varnames:
    - FeatureTypeStatic
    - FeatureTypeMetered
  models.HttpResponse-array_models_Feature:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Feature'
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.HttpResponse-models_Feature:
    properties:
      data:
        $ref: '#/definitions/models.Feature'
      message:
        type: string
      status:
        type: integer
    type: object
  models.HttpResponse-models_Meter:
    properties:
      data:
        $ref: '#/definitions/models.Meter'
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.HttpResponse-pagination_PaginationView-models_Meter:
    properties:
      data:
        $ref: '#/definitions/pagination.PaginationView-models_Meter'
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.Meter:
    properties:
      aggregation:
        $ref: '#/definitions/models.AggregationEnum'
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      event_type:
        type: string
      id:
        type: string
      name:
        type: string
      properties:
        items:
          type: string
        type: array
      slug:
        type: string
      tenant_slug:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      value_property:
        type: string
    type: object
...
  pagination.PaginationView-models_Meter:
    properties:
      limit:
        type: integer
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.Meter'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /v1/features:
    get:
      consumes:
      - application/json
      description: Get a list of features for the tenant
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - description: Filter by feature type (static/metered)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Features retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-array_models_Feature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List features
      tags:
      - features
    post:
      consumes:
      - application/json
      description: Create a new feature for the tenant
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Feature creation data
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/features.createFeatureRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Feature created successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_Feature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new feature
      tags:
      - features
  /v1/features/{idOrSlug}:
    delete:
      consumes:
      - application/json
      description: Delete a feature by ID or slug
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Feature ID or Slug
        in: path
        name: idOrSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Feature deleted successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Feature not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete a feature
      tags:
      - features
    get:
      consumes:
      - application/json
      description: Get feature details by ID or slug
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Feature ID or Slug
        in: path
        name: idOrSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Feature retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_Feature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Feature not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get feature details
      tags:
      - features
    put:
      consumes:
      - application/json
      description: Update a feature by ID or slug
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Feature ID or Slug
        in: path
        name: idOrSlug
        required: true
        type: string
      - description: Feature update data
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/features.updateFeatureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Feature updated successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_Feature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Feature not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a feature
      tags:
      - features
  /v1/meters:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all meters for the tenant
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Meters retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-pagination_PaginationView-models_Meter'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List all meters
      tags:
      - meters
    post:
      consumes:
      - application/json
      description: Create a new meter for the tenant
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Meter creation data
        in: body
        name: meter
        required: true
        schema:
          $ref: '#/definitions/meters.createMeterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Meter created successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_Meter'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new meter
      tags:
      - meters
  /v1/meters/{idOrSlug}:
    delete:
      consumes:
      - application/json
      description: Delete a meter by ID or slug
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Meter ID or slug
        in: path
        name: idOrSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Meter deleted successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Meter not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete a meter
      tags:
      - meters
    get:
      consumes:
      - application/json
      description: Get details of a specific meter by ID or slug
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Meter ID or slug
        in: path
        name: idOrSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meter retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_Meter'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Meter not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get meter details
      tags:
      - meters
    put:
      consumes:
      - application/json
      description: Update a meter's details by ID or slug
      parameters:
      - description: Tenant Slug
        in: header
        name: X-Tenant-Slug
        required: true
        type: string
      - description: Meter ID or slug
        in: path
        name: idOrSlug
        required: true
        type: string
      - description: Meter update data
        in: body
        name: meter
        required: true
        schema:
          $ref: '#/definitions/meters.updateMeterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Meter updated successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_Meter'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Meter not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a meter
      tags:
      - meters
...
swagger: "2.0"